[Fix] 모드선택
모드선택을 룸패널로 이동
