+++ Assets/HwangGyuHo/HGH_Scene.unity
@@ -190,17 +190,44 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1 &541217146 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 4013794221103789331, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+  m_PrefabInstance: {fileID: 4013794220077840773}
+  m_PrefabAsset: {fileID: 0}
 --- !u!114 &1059671194 stripped
 MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 4013794221103789343, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
   m_PrefabInstance: {fileID: 4013794220077840773}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
+  m_GameObject: {fileID: 541217146}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 66707ee8ed604824f8390efc2f07e745, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &1059671198
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 541217146}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: aa584fbee541324448dd18d8409c7a41, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObservedComponentsFoldoutOpen: 1
+  Group: 0
+  prefixField: -1
+  Synchronization: 3
+  OwnershipTransfer: 0
+  observableSearch: 2
+  ObservedComponents: []
+  sceneViewId: 4
+  InstantiationId: 0
+  isRuntimeInstantiated: 0
 --- !u!1 &1154959252 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 4013794219872418478, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
@@ -444,6 +471,26 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 604240684824781010, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+      propertyPath: m_Colors.m_HighlightedColor.b
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 604240684824781010, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+      propertyPath: m_Colors.m_HighlightedColor.g
+      value: 0.620383
+      objectReference: {fileID: 0}
+    - target: {fileID: 604240684824781010, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+      propertyPath: m_Colors.m_HighlightedColor.r
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 604240684824781010, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
+      value: SendSelectMode3
+      objectReference: {fileID: 0}
+    - target: {fileID: 3420512900761012582, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
+      value: SendSelectMode2
+      objectReference: {fileID: 0}
     - target: {fileID: 3518283138475548969, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
       propertyPath: m_RootOrder
       value: 7
@@ -660,6 +707,10 @@ PrefabInstance:
       propertyPath: m_IsActive
       value: 1
       objectReference: {fileID: 0}
+    - target: {fileID: 4013794221103789343, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+      propertyPath: photonView
+      value: 
+      objectReference: {fileID: 1059671198}
     - target: {fileID: 4013794221106822146, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
       propertyPath: m_Name
       value: Canvas
@@ -747,6 +798,14 @@ PrefabInstance:
     - target: {fileID: 4013794221106822158, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6575730037212357435, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+      propertyPath: m_Interactable
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6575730037212357435, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
+      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
+      value: SendSelectMode1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 4031c796270f65847b7a655eda245bf2, type: 3}
