--- Assets/HwangGyuHo/HGH_Scripts/DragBugSafe.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3db1d455c0080ec4b931e81d6a550802
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
